# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Library Management API",
    "description": "## Library Management API\n\nA full-featured REST API for managing a library system including:\n\n- **Authentication** \u2013 Register, login (JWT Bearer), and logout\n- **Users** \u2013 Admin CRUD for user accounts\n- **Library Branches** \u2013 Manage physical branch locations\n- **Books** \u2013 Catalog management with search & filtering\n- **Loans** \u2013 Borrow/return workflow with status tracking\n\n### Authentication\nMost endpoints require a **Bearer JWT token**. Obtain one via `POST /api/v1/auth/login` (OAuth2 password flow) or `POST /api/v1/auth/register`.\n\n### Roles\n| Role | Description |\n|------|-------------|\n| `member` | Default role \u2013 can browse books and request loans |\n| `librarian` | Can manage books and branches, approve/reject loans |\n| `admin` | Full access including user management |\n",
    "contact": {
      "name": "Library API Support",
      "email": "support@library.com"
    },
    "license": {
      "name": "MIT"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local development"
    }
  ],
  "paths": {
...
    "/api/v1/loans": {
      "post": {
        "tags": [
          "Loans"
        ],
        "summary": "Request a new loan",
        "description": "Request to borrow a book from a specific branch. Any authenticated user can request a loan.",
        "operationId": "create_loan_endpoint_api_v1_loans_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoanCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Loan request created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoanResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid data (e.g. book not available)"
          },
          "401": {
            "description": "Not authenticated"
          },
          "422": {
            "description": "Validation error"
          }
        }
      },
      "get": {
        "tags": [
          "Loans"
        ],
        "summary": "List loans",
        "description": "Retrieve a paginated list of loans. Members see only their own; Librarians and Admins see all loans with optional filters.",
        "operationId": "list_loans_api_v1_loans_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Size"
            }
          },
          {
            "name": "member_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Member Id"
            }
          },
          {
            "name": "branch_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Branch Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/LoanStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "sort_by",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "created_at",
              "title": "Sort By"
            }
          },
          {
            "name": "sort_order",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(asc|desc)$",
              "default": "desc",
              "title": "Sort Order"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of loans",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoanListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/loans/my-history": {
      "get": {
        "tags": [
          "Loans"
        ],
        "summary": "My loan history",
        "description": "Retrieve the current authenticated user's loan history with optional status filter.",
        "operationId": "my_loan_history_api_v1_loans_my_history_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Size"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/LoanStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "sort_by",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "created_at",
              "title": "Sort By"
            }
          },
          {
            "name": "sort_order",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(asc|desc)$",
              "default": "desc",
              "title": "Sort Order"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of user's loans",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoanListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/loans/{loan_id}": {
      "get": {
        "tags": [
          "Loans"
        ],
        "summary": "Get loan details",
        "description": "Retrieve a single loan by its ID. Members can only view their own loans.",
        "operationId": "get_loan_endpoint_api_v1_loans__loan_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "loan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Loan Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Loan details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoanResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Access denied (member viewing another user's loan)"
          },
          "404": {
            "description": "Loan not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/loans/{loan_id}/status": {
      "patch": {
        "tags": [
          "Loans"
        ],
        "summary": "Update loan status",
        "description": "Transition the loan to a new status. Allowed transitions depend on the user's role:\n\n- **Member**: can cancel their own `requested` loan\n- **Librarian**: can approve, reject, mark as borrowed/returned\n- **Admin**: full control over all status transitions\n\nStatus flow: `requested` \u2192 `approved` \u2192 `borrowed` \u2192 `returned` (or `overdue` / `lost`)",
        "operationId": "update_loan_status_endpoint_api_v1_loans__loan_id__status_patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "loan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Loan Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoanStatusUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Loan status updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoanResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid status transition"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Insufficient permissions for this transition"
          },
          "422": {
            "description": "Validation error"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LoanCreate": {
        "properties": {
          "book_id": {
            "type": "string",
            "title": "Book Id"
          },
          "branch_id": {
            "type": "string",
            "title": "Branch Id"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "book_id",
          "branch_id"
        ],
        "title": "LoanCreate"
      },
      "LoanListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/LoanResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "size": {
            "type": "integer",
            "title": "Size"
          },
          "pages": {
            "type": "integer",
            "title": "Pages"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "size",
          "pages"
        ],
        "title": "LoanListResponse"
      },
      "LoanResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "member_id": {
            "type": "string",
            "title": "Member Id"
          },
          "book_id": {
            "type": "string",
            "title": "Book Id"
          },
          "branch_id": {
            "type": "string",
            "title": "Branch Id"
          },
          "borrow_date": {
            "type": "string",
            "format": "date-time",
            "title": "Borrow Date"
          },
          "due_date": {
            "type": "string",
            "format": "date-time",
            "title": "Due Date"
          },
          "return_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Return Date"
          },
          "status": {
            "$ref": "#/components/schemas/LoanStatus"
          },
          "late_fee": {
            "type": "number",
            "title": "Late Fee"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "status_history": {
            "items": {
              "$ref": "#/components/schemas/LoanStatusHistoryResponse"
            },
            "type": "array",
            "title": "Status History",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "member_id",
          "book_id",
          "branch_id",
          "borrow_date",
          "due_date",
          "return_date",
          "status",
          "late_fee",
          "notes",
          "created_at",
          "updated_at"
        ],
        "title": "LoanResponse"
      },
      "LoanStatus": {
        "type": "string",
        "enum": [
          "requested",
          "canceled",
          "approved",
          "borrowed",
          "overdue",
          "returned",
          "lost"
        ],
        "title": "LoanStatus"
      },
      "LoanStatusHistoryResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "previous_status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LoanStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "new_status": {
            "$ref": "#/components/schemas/LoanStatus"
          },
          "changed_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Changed By"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "changed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Changed At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "previous_status",
          "new_status",
          "changed_by",
          "notes",
          "changed_at"
        ],
        "title": "LoanStatusHistoryResponse"
      },
      "LoanStatusUpdate": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/LoanStatus"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "LoanStatusUpdate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  },
...
}